requirements.txt
PyQt5>=5.15,<6.6
PyOpenGL>=3.1,<3.2
numpy==1.26.4
trimesh>=4.0,<4.2
numpy-stl>=3.0,<3.2
shapely>=2.0,<2.1  # optional; kontur birleştirme için kullanılır, yoksa convex hull fallback var

# Opsiyonel (demo/test)
pyqtgraph>=0.13,<0.14
open3d>=0.18,<0.19
